'feat: Add function to return tuples of names and gtfs_ids'
I think I will need to store both the names along with the gtfs_id
of each station.

When selecting the station in selectStation(line) in main.py,
the code currently posts the station names but this isn't
enough to specify the station for later use because some stations
have the same name.

I need to get the gtfs_id from selecStation(line) so that I can
store it when making a SavedStation object in the database.