"feat: Revert times_dict_from_gtfs_array function to return GTFS ids instead of station names"
I realized that I would need both the GTFS ids and station names
later on. It is better to store the gtfs ids because there are no
duplicate gtfs ids but there are duplicate station names.

It is easier to get the station name from the gtfs id because I can
simply look up the station name of the given id but I can't always
get the gtfs id from the station name without additional information.